+++ src/pages/SupportGroups/SupportGroups.jsx
@@ -6,6 +6,7 @@ const SupportGroups = ({ value }) => {
   return (
     <SupportGroupsWrapper>
       <p>{value}</p>
+      <h1>This is the Support Group page. It is here to check the Router</h1>
     </SupportGroupsWrapper>
   );
 };
