--- src/pages/TandC/index.js
@@ -1,3 +1,0 @@
-import TermsAndConditions from './TandC';
-
-export default TermsAndConditions;
