+++ src/pages/Terms/Terms.jsx
@@ -1,12 +1,12 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { TandCWrapper } from './TandC.styles';
+import { TermsWrapper } from './Terms.styles';
 
 const TermsAndConditions = ({ value }) => {
   return (
-    <TandCWrapper>
+    <TermsWrapper>
       <p>{value}</p>
-    </TandCWrapper>
+    </TermsWrapper>
   );
 };
