+++ src/pages/Terms/Terms.styles.js
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 
-export const TandCWrapper = styled.div`
+export const TermsWrapper = styled.div`
   padding: var(--small-space);
   border: 0.1rem solid var(--logopink);
 `;
