+++ src/pages/Terms/Terms.test.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render, screen } from '@testing-library/react';
 
-import TermsAndConditions from './TandC';
+import TermsAndConditions from './Terms';
 
 describe('TermsAndConditions', () => {
   test('renders TermsAndConditions component', () => {
