+++ src/pages/Terms/index.js
@@ -1,0 +1,3 @@
+import TermsAndConditions from './Terms';
+
+export default TermsAndConditions;
